:art: fix: improved folder structure
introduce router, barrel export and fix structure
