+++ .env.example
@@ -1,3 +1,2 @@
-API_URL=<API URL>
 MONGO_URL=<Mongodb connection url>
 PORT=<Preferred server port>
